Fix with new changes
